Adding Wacom removal "tools"
Since the Wacom Bamboo and Intuos/Cintiq drivers seems to not play well
while being on the same machine we usually remove the driver that is
not currently being installed.  These scripts check if the apps are
installed and if they are they run the uninstall scripts that are built
into the apps.