New Script
This script will kill all browsers in our environment if the user has
been idle for 10 minutes. (I wrote this because we have people
complaining about the browsers keeping their accounts logged in which
is actually a problem of users not quitting browsers when done. It just
checks idle time and if a browser is running and kills it.) I run this
at every check-in.