Initial upload (v0.8)
This script was originally written by Calum Hunter to detect internal
drives and format them to a single volume.  It also re-joins broken
fusion drives.  We use this to recreate the fusion drives and to
re-partition our drives to a single volume.

I only added the “CHECK_FOR_APFS” function which allows the script to
remove APFS containers to quickly go back to 10.12 from 10.13 with APFS.

We drop this on the desktop of out NetBoot set.